refactor: move queryset from drivers to types
