+++ jina/types/querylang/queryset/helper.py
@@ -1,7 +1,7 @@
 from functools import partial
 from typing import Callable, List, Type, Union, Iterable
 
-from ....excepts import LookupyError
+from jina.excepts import LookupyError
 
 
 def iff(precond: Callable, val: Union[int, str], f: Callable) -> bool:
